backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
tensorflow==2.20.0
pillow>=10.4.0
requests==2.31.0
numpy>=1.26.0


